confidence graph folder rename
