+++ host_v2/src/components/ConfidenceGraph/ConfidenceResponseLegend.tsx
