+++ host_v2/src/components/ConfidenceGraph/CustomBar.tsx
